+++ chart/src/main/java/cn/jingzhuan/lib/chart3/drawline/RayDrawLine.kt
@@ -45,10 +45,10 @@ class RayDrawLine<T : AbstractDataSet<*>>(chart: AbstractChartView<T>) : Abstrac
         // 没有吸附并且没有抬起时平顺滑动
         if (!chartView.isDrawLineAdsorb && !dataSet.isActionUp) {
             val startX = startPoint.x
-            val startY = chartView.getScaleY(startPoint.value, lMax, lMin)
+            val startY = chartView.getScaleY(startPoint.value.toFloat(), lMax, lMin)
 
             val endX = endPoint.x
-            val endY = chartView.getScaleY(endPoint.value, lMax, lMin)
+            val endY = chartView.getScaleY(endPoint.value.toFloat(), lMax, lMin)
 
             drawShape(canvas, dataSet, startX, startY, endX, endY)
             return
@@ -56,14 +56,14 @@ class RayDrawLine<T : AbstractDataSet<*>>(chart: AbstractChartView<T>) : Abstrac
 
         val startIndex = getIndexInTime(dataSet, baseDataSet, startPoint.time)
         val startX = getEntryX(startIndex, baseDataSet) ?: return
-        val startY = chartView.getScaleY(startPoint.value, lMax, lMin)
+        val startY = chartView.getScaleY(startPoint.value.toFloat(), lMax, lMin)
         if (!dataSet.isSelect || dataSet.isActionUp) {
             startPoint.apply { dataIndex = startIndex; x = startX; y = startY }
         }
 
         val endIndex = getIndexInTime(dataSet, baseDataSet, endPoint.time)
         val endX = getEntryX(endIndex, baseDataSet) ?: return
-        val endY = chartView.getScaleY(endPoint.value, lMax, lMin)
+        val endY = chartView.getScaleY(endPoint.value.toFloat(), lMax, lMin)
         if (!dataSet.isSelect || dataSet.isActionUp) {
             endPoint.apply { dataIndex = endIndex; x = endX; y = endY }
         }
